requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
scipy>=1.11.0
PyYAML>=6.0
matplotlib>=3.7.0
Pillow>=10.0.0
lpips>=0.1.4
pytorch-msssim>=1.0.0
pytest>=7.4.0
tqdm>=4.66.0
